design GaussianSplatGroup around BatchedMesh paradigm, simplier / cleaner.
